refactor(editor): Phase 4 - useEditorEffects 훅 추출
- useEditorEffects.ts: 초기화 및 자동 선택 로직 (~90줄) 분리
- EditorPage.tsx: 관련 이펙트 제거, 훅 호출로 대체
- CC: 20 (목표 미달, 추가 컴포넌트 분리 필요)
